+++ src/com/emc/paradb/advisor/ui/transaction_distribution/TransactionDistributionCB.java
@@ -6,5 +6,5 @@ import java.util.Map;
 public interface TransactionDistributionCB
 {
 	public void draw(int dist, int nonDist, Map<Integer, Integer> nodeAccess);
-	public void draw(HashMap<String, Float> tdMap);
+	public void draw(HashMap<String, Integer> tdMap);
 }
